precision of the camera motion improved
- changed Camera::proceedControl to take time argument as float in
seconds instead of integer in milliseconds
- made changes in windows version to use QueryPerformanceCounter instead
of GetTickCount for more precise camera control
- made changes in linux version to match new Camera::proceedControl time
argument

+++ src/common/Render.cpp
@@ -56,7 +56,7 @@ void Render::setImageSize(unsigned int width, unsigned int height)
 
   if(width > 0 && height > 0)
   {
-    if(width * height > (int)image.size())
+    if(width * height > image.size())
       image.resize(width * height);
 
     Color * pCol = &image.front();
